create admin office

+++ src/main/java/com/mahait/gov/in/repository/UserInfoRepoImpl.java
@@ -112,7 +112,7 @@ public class UserInfoRepoImpl implements UserInfoRepo {
 	
 
 	@Override
-	public OrgUserMst getUserByUserId(int updatedUserId) {
+	public OrgUserMst getUserByUserId(Long updatedUserId) {
 		Session currentSession = entityManager.unwrap(Session.class);
 		return currentSession.find(OrgUserMst.class, updatedUserId);
 	}
@@ -122,7 +122,7 @@ public class UserInfoRepoImpl implements UserInfoRepo {
 		Session currentSession = entityManager.unwrap(Session.class);
 	    return 	(OrgUserMst) currentSession.save(orgUserMst);
 	}
- 
+
